pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "endstone-primebds"
version = "3.2.0"
description = "An essentials plugin for diagnostics, stability, and quality of life on Minecraft Bedrock Edition."
author = "PrimeStrat"
dependencies = [
    "bedrock-protocol-packets==0.0.7",
    "mcbe-binarystream==2.1.0",
    "chest_form_api_endstone >= 1.0.0"
]

[project.entry-points."endstone"]
primebds = "endstone_primebds:PrimeBDS"

